+++ packages/page-navigation/src/types/StoreProps.ts
@@ -6,11 +6,26 @@
  * @copyright 2019-2022 Nguyen Huu Phuoc <me@phuoc.ng>
  */
 
-import type { PdfJs } from '@react-pdf-viewer/core';
+import type { DestinationOffsetFromViewport, PdfJs } from '@react-pdf-viewer/core';
+import { SpecialZoomLevel } from '@react-pdf-viewer/core';
+
+export interface JumpFromAnnotation {
+    bottomOffset: number;
+    dest: string;
+    leftOffset: number;
+    pageIndex: number;
+}
 
 export interface StoreProps {
     currentPage?: number;
     doc?: PdfJs.PdfDocument;
-    numberOfPages?: number;
+    jumpFromAnnotation?: JumpFromAnnotation;
+    jumpToDestination?(
+        pageIndex: number,
+        bottomOffset: number | DestinationOffsetFromViewport,
+        leftOffset: number | DestinationOffsetFromViewport,
+        scaleTo?: number | SpecialZoomLevel
+    ): void;
     jumpToPage?(pageIndex: number): void;
+    numberOfPages?: number;
 }