+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ## v3.7.0 [WIP]
 
-**Bug fix**
+**Bug fixes**
 
 -   The page navigation options are missing when creating a toolbar plugin
+-   The search popover always shows up after pressing shortcuts
 
 ## v3.6.0