+++ CHANGELOG.md
@@ -46,6 +46,7 @@ const defaultLayoutPluginInstance = defaultLayoutPlugin({
 -   Clicking a particular bookmark might not go to the destination
 -   Link annotations with unsafe URLs won't be rendered to avoid secutiry issues
 -   The `CharacterMap` type isn't available
+-   The `onPageChange` callback does not trigger if the current page equals to the initial page
 -   The page navigation options are missing when creating a toolbar plugin
 -   The search popover always shows up after pressing shortcuts
 -   The viewer always navigates to previous, next pages after users press shortcuts even the document isn't focused